R4efactoring: Rename to match contents
