+++ leitus/session.py
@@ -24,7 +24,7 @@ import os
 import os.path
 from tempfile import NamedTemporaryFile
 
-from leitus.deep import CryptDeviceWithRandomKey
+from leitus.crypt import CryptDeviceWithRandomKey
 from leitus.filesystem import ExtFileSystem
 from leitus.loop import LoopDevice
